chore(watermark): split demo
