+++ src/packages/__VUE/watermark/demo/multi.vue
@@ -1,0 +1,10 @@
+<template>
+  <nut-cell style="height: 240px">
+    <nut-watermark
+      :full-page="false"
+      :gap-y="24"
+      font-color="#fa2c19"
+      :content="['nutui', 'watermark']"
+    ></nut-watermark>
+  </nut-cell>
+</template>
