+++ src/packages/__VUE/watermark/doc.en-US.md
@@ -14,116 +14,17 @@ app.use(Watermark);
 
 ### Basic Usage
 
-:::demo
-
-```vue
-<template>
-  <nut-cell class="wrap">
-    <nut-button @click="showTextMark">Text Watermark</nut-button>
-    <nut-button @click="showImageMark">Image Watermark</nut-button>
-    <nut-watermark v-if="!flag" class="mark1" :z-index="1" content="nut-ui-water-mark"></nut-watermark>
-    <nut-watermark
-      v-if="flag"
-      class="mark1"
-      :image-width="60"
-      :image-height="23"
-      :z-index="1"
-      :image="imgSrc"
-    ></nut-watermark>
-  </nut-cell>
-</template>
-<script setup>
-import { ref } from 'vue';
-const imgSrc = ref('//img11.360buyimg.com/imagetools/jfs/t1/57345/6/20069/8019/62b995cdEd96fef03/51d3302dfeccd1d2.png');
-const flag = ref(false);
-const showTextMark = () => {
-  flag.value = false;
-};
-const showImageMark = () => {
-  flag.value = true;
-};
-</script>
-<style>
-.wrap {
-  width: 100%;
-  height: 240px;
-  display: block;
-  background: #fff;
-}
-.wrap > img {
-  width: 100%;
-}
-.mark1 {
-  width: 100%;
-  top: 50px;
-}
-.nut-button {
-  margin-right: 10px;
-}
-</style>
-```
-
-:::
+> demo: watermark basic
 
 ### Part Usage
 
-:::demo
+> demo: watermark part
 
-```vue
-<template>
-  <nut-cell class="wrap">
-    <img :src="src" alt="" />
-    <nut-watermark :fullPage="false" font-color="#fa2c19" content="nut-ui"></nut-watermark>
-  </nut-cell>
-</template>
-<script setup>
-import { ref } from 'vue';
-const src = ref('//img10.360buyimg.com/ling/jfs/t1/181258/24/10385/53029/60d04978Ef21f2d42/92baeb21f907cd24.jpg');
-</script>
-<style>
-.wrap {
-  width: 100%;
-  height: 240px;
-  display: block;
-  background: #fff;
-}
-.wrap > img {
-  width: 100%;
-}
-</style>
-```
-
-:::
-
-### Multiline Watermark
+### Multiline Watermark v4.2.5
 
 The content attribute is passed into a string array, showing a multi-line watermark
 
-:::demo
-
-```vue
-<template>
-  <nut-cell class="wrap">
-    <nut-watermark
-      :full-page="false"
-      :gap-y="24"
-      font-color="#fa2c19"
-      :content="['nut-ui', 'water-mark']"
-    ></nut-watermark>
-  </nut-cell>
-</template>
-<script setup></script>
-<style>
-.wrap {
-  width: 100%;
-  height: 240px;
-  display: block;
-  background: #fff;
-}
-</style>
-```
-
-:::
+> demo: watermark multi
 
 ## API
 
@@ -138,7 +39,7 @@ The content attribute is passed into a string array, showing a multi-line waterm
 | image-width | Width of image | number | `120` |
 | image-height | Height of image | number | `64` |
 | z-index | Z-index of the appended watermark element | number | `2000` |
-| content | Watermark text content | string | string[] | - |
+| content | Watermark text content | string \| string[] | - |
 | font-color | Watermark text color | string | `rgba(0, 0, 0, .15)` |
 | font-size | Watermark text font size | string \| number | `16` |
 | gap-x | Horizontal spacing between watermarks | number | `24` |
